FARRAGO:  (FRG-270) fix a couple of issues caught by findbugs
[git-p4: depot-paths = "//open/lu/dev/": change = 9287]

+++ farrago/src/org/eigenbase/jmi/JmiChangeSet.java
@@ -756,7 +756,7 @@ public class JmiChangeSet
 
                 // Avoid unnecessary updates in the case where ordinals
                 // are already correct.
-                if (oldOrdinal != newOrdinal) {
+                if (oldOrdinal.intValue() != newOrdinal.intValue()) {
                     target.refSetValue("ordinal", newOrdinal);
                 }
             }
